# api-docs/swagger.yaml
openapi: 3.0.0
info:
  title: Lessons Marketplace API
  description: API for the Lessons Marketplace application
  version: 1.0.0
servers:
  - url: 'http://localhost:3000/api/v1'
    description: Development server
  - url: '/api/v1'
    description: Production relative path
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LessonType:
      type: string
      enum: [VOICE, GUITAR, BASS, DRUMS]
    GoalStatusValue:
      type: string
      enum: [CREATED, IN_PROGRESS, COMPLETED, CANCELLED]
      description: Possible statuses for a goal
...
    ObjectiveStatusValue:
      type: string
      enum: [CREATED, IN_PROGRESS, ACHIEVED, ABANDONED]
      description: Possible statuses for a student objective
    ObjectiveStatus:
      type: object
      properties:
        id:
          type: string
          format: uuid
        objectiveId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/ObjectiveStatusValue'
        context:
          type: object
          nullable: true
          description: Optional context for the status change (JSON)
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - objectiveId
        - status
        - createdAt
    Address:
      type: object
      properties:
        id:
          type: string
        street:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - street
        - city
        - state
        - postalCode
        - country
    Student:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        dateOfBirth:
          type: string
          format: date
      required:
        - id
        - firstName
        - lastName
        - email
    Teacher:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        dateOfBirth:
          type: string
          format: date
      required:
        - id
        - firstName
        - lastName
        - email
    TeacherWithRates:
      allOf:
        - $ref: '#/components/schemas/Teacher'
        - type: object
          properties:
            lessonHourlyRates:
              type: object
              additionalProperties:
                type: number
    Goal:
      type: object
      properties:
        id:
          type: string
        lessonId:
          type: string
        title:
          type: string
        description:
          type: string
        estimatedLessonCount:
          type: integer
          minimum: 1
        currentStatus:
          $ref: '#/components/schemas/GoalStatusValue'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - lessonId
        - title
        - description
        - estimatedLessonCount
    LessonRequest:
      type: object
      properties:
        id:
          type: string
        type:
          $ref: '#/components/schemas/LessonType'
        startTime:
          type: string
          format: date-time
        durationMinutes:
          type: integer
          minimum: 1
        address:
          $ref: '#/components/schemas/Address'
        studentId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - type
        - startTime
        - durationMinutes
        - address
        - studentId
...
    Lesson:
      type: object
      properties:
        id:
          type: string
        quoteId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        currentStatusId:
          type: string
        goals:
          type: array
          items:
            $ref: '#/components/schemas/Goal'
      required:
        - quoteId
        - createdAt
        - updatedAt
        - currentStatusId
...
    Objective:
      type: object
      properties:
        id:
          type: string
          format: uuid
        studentId:
          type: string
          format: uuid
        lessonType:
          $ref: '#/components/schemas/LessonType'
        title:
          type: string
        description:
          type: string
        targetDate:
          type: string
          format: date-time
        currentStatusId:
          type: string
          format: uuid
          nullable: true
        currentStatus:
          $ref: '#/components/schemas/ObjectiveStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - studentId
        - lessonType
        - title
        - description
        - targetDate
        - currentStatus
        - createdAt
        - updatedAt
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
security:
  - BearerAuth: []
paths:
...
  /auth/refresh-token:
    post:
      summary: Refresh an expired access token
      tags:
        - Auth
      security: []
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
        '401':
          description: Unauthorized or invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /goals:
    post:
      summary: Create a new goal
      tags:
        - Goals
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                lessonId:
                  type: string
                title:
                  type: string
                description:
                  type: string
                estimatedLessonCount:
                  type: integer
                  minimum: 1
              required:
                - lessonId
                - title
                - description
                - estimatedLessonCount
      responses:
        '201':
          description: Goal created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /lessons:
    post:
      summary: Create a new lesson from a quote
      tags:
        - Lessons
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                quoteId:
                  type: string
              required:
                - quoteId
      responses:
        '201':
          description: Lesson created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lesson'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /lesson-requests:
    post:
      summary: Create a new lesson request
      tags:
        - Lesson Requests
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  $ref: '#/components/schemas/LessonType'
                startTime:
                  type: string
                  format: date-time
                durationMinutes:
                  type: integer
                  minimum: 1
                addressObj:
                  $ref: '#/components/schemas/Address'
                studentId:
                  type: string
              required:
                - type
                - startTime
                - durationMinutes
                - addressObj
                - studentId
      responses:
        '201':
          description: Lesson request created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LessonRequest'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /students:
    post:
      summary: Create a new student
      tags:
        - Students
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                  example: 'Test'
                lastName:
                  type: string
                  example: 'Student'
                email:
                  type: string
                  format: email
                  example: 'test.student.new@example.com'
                password:
                  type: string
                  format: password
                  example: 'password123'
                phoneNumber:
                  type: string
                  example: '111-222-3333'
                dateOfBirth:
                  type: string
                  format: date
                  example: '2005-06-07'
              required:
                - firstName
                - lastName
                - email
                - password
                - phoneNumber
                - dateOfBirth
      responses:
        '201':
          description: Student created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /teachers:
    get:
      summary: Get available teachers
      tags:
        - Teachers
      parameters:
        - name: lessonType
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/LessonType'
        - name: limit
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            default: 5
      responses:
        '200':
          description: List of available teachers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeacherWithRates'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /objectives:
    get:
      summary: Get objectives for the authenticated student
      tags:
        - Objectives
      security:
        - BearerAuth: []
      # Add query parameters if filtering is implemented (e.g., status, lessonType)
      # parameters:
      #   - name: status
      #     in: query
      #     required: false
      #     schema:
      #       $ref: '#/components/schemas/ObjectiveStatusValue'
      #     description: Filter objectives by status
      responses:
        '200':
          description: A list of objectives for the authenticated student
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Objective'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a new objective for the authenticated student
      tags:
        - Objectives
      security:
        - BearerAuth: []
      # Removed studentId from path parameter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                # studentId is no longer in the body, derived from authenticated user
                title:
                  type: string
                  example: "Master the C Major Scale"
                description:
                  type: string
                  example: "Be able to play the C Major scale ascending and descending at 80 bpm."
                lessonType:
                  $ref: '#/components/schemas/LessonType'
                  example: GUITAR
                targetDate:
                  type: string
                  format: date-time
                  example: "2024-12-31T23:59:59Z"
              required:
                - title
                - description
                - lessonType
                - targetDate
      responses:
        '201':
          description: Objective created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Objective'
        '400':
          description: Bad request (e.g., missing fields, invalid data)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...